refactor(validators): restructure user validation methods
Restructure namespaces and refactor validation methods in
UserService and UserServiceValidator classes. The validators
namespace has been updated to `HoneyHub.Users.AppService.Services.Validators.Users`
for better organization. Introduced helper methods for
validating user fields, password requirements, and
username/email uniqueness to enhance code readability and
maintainability. Updated validation logic for external and
admin user requests to ensure consistency.

BREAKING CHANGE: Updated namespace for validators.
